pyproject.toml
[project]
name = "super-mario-bros-ppo-pytorch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gym-super-mario-bros>=7.4.0",
    "gymnasium>=1.2.2",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "torch>=2.9.1",
]
